appeareance changes pt.1

+++ StrokeAnalyzer/app/src/main/java/com/example/asus/strokeanalyzer/Entities/Patient.java
@@ -22,5 +22,5 @@ public class Patient {
     public String Surname;
 
     @ColumnInfo(name = "patient_name")
-    public int PatientNumber;
+    public long PatientNumber;
 }
